app.py
from flask import Flask, request, jsonify

app = Flask(__name__)

# Simple route
@app.route("/")
def home():
    return "Welcome to the Flask app!"

# Login route
@app.route("/login", methods=["POST"])
def login():
    data = request.get_json()
    username = data.get("username")
    password = data.get("password")
    if username == "admin" and password == "password":
        return jsonify({"message": "Login successful!"}), 200
    else:
        return jsonify({"message": "Invalid credentials"}), 401

if __name__ == "__main__":
    app.run(debug=True)
